% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randgeo-package.R
\docType{package}
\name{randgeo-package}
\alias{randgeo}
\alias{randgeo-package}
\title{Random WKT or GeoJSON}
\description{
Random WKT or GeoJSON
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

