% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_point.R
\name{rg_point}
\alias{rg_point}
\title{Random point}
\usage{
rg_point(count = 1, bbox = NULL)
}
\arguments{
\item{count}{(integer/numeric) number of Polygons. Default: 1}

\item{bbox}{(integer/numeric) bounding box, numeric vector of the form
\code{west, south, east, north}. optional}
}
\value{
GeoJSON; a list with one ore more Points in a FeatureCollection
}
\description{
Random point
}
\examples{
rg_point()
rg_point(10)
rg_point(bbox = c(50, 50, 60, 60))
}

