% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg_position.R
\name{rg_position}
\alias{rg_position}
\title{Random position}
\usage{
rg_position(bbox = NULL)
}
\arguments{
\item{bbox}{(integer/numeric) bounding box, numeric vector of the form
\code{west, south, east, north}. optional}
}
\value{
numeric vector of long, lat
}
\description{
Random position
}
\examples{
rg_position()
rg_position(bbox = c(50, 50, 60, 60))
}
