% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_polygon.R
\name{geo_polygon}
\alias{geo_polygon}
\title{Random GeoJSON polygon}
\usage{
geo_polygon(count = 1, num_vertices = 10, max_radial_length = 10,
  bbox = NULL)
}
\arguments{
\item{count}{(integer/numeric) number of Polygons. Default: 1}

\item{num_vertices}{(integer/numeric) is default 10 and is how many
coordinates each Polygon will contain. Default: 10}

\item{max_radial_length}{(integer/numeric) maximum number of decimal degrees
latitude or longitude that a vertex can reach out of the center of the
Polygon. Default: 10}

\item{bbox}{(integer/numeric) bounding box, numeric vector of the form
\code{west, south, east, north}. optional}
}
\value{
GeoJSON; a list with one ore more Polygons in a FeatureCollection
}
\description{
Random GeoJSON polygon
}
\examples{
geo_polygon()
geo_polygon(10)
geo_polygon(bbox = c(50, 50, 60, 60))
}
