% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randgeo-package.R
\docType{package}
\name{randgeo-package}
\alias{randgeo-package}
\alias{randgeo}
\title{Random WKT or GeoJSON}
\description{
Adapted from \url{https://github.com/tmcw/geojson-random} for a pure
R implementation
}
\section{Package API}{

\itemize{
 \item rg_position - random position
 \item geo_point - random GeoJSON point
 \item geo_polygon - random GeoJSON polygon
 \item wkt_point - random WKT point
 \item wkt_polygon - random WKT polygon
}
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Noam Ross
}
\keyword{package}
