% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randgeo-package.R
\docType{package}
\name{randgeo-package}
\alias{randgeo-package}
\alias{randgeo}
\title{Random WKT or GeoJSON}
\description{
adapted from \url{https://github.com/tmcw/geojson-random} to have a pure
R implementation
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
